Created a simple image class and began working on the load function.
So far the function reads through a well formed header and loads a single pixel. Next up is to read till the end of the file. In the future this
function should be made safe and sane by validating input instead of blindly trusting that the ppm files are well formed, but for the project I need
this for--this week--all of the ppm files *are* well formed.